backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
pypdf2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
redis==5.0.1
apscheduler==3.10.4
celery==5.4.0
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.2
